Add files via upload
Potential update to include Changelly API to facilitate exchanges to 150+ cryptocurrencies. Morality Co-operative is undergoing authorisation with FCA (which is required to facilitate exchanges of this nature) but could be integrated now under The Co-op's temporary authorisation status. 

Ethereum Gas Fees currently making this feature almost redundant and something to revisit in the future.